+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/resttemplate/fluent/HttpMethodBuilder.groovy
@@ -1,5 +1,6 @@
 package com.ofg.infrastructure.web.resttemplate.fluent
 
+import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.delete.DeleteMethod
 import com.ofg.infrastructure.web.resttemplate.fluent.delete.DeleteMethodBuilder
 import com.ofg.infrastructure.web.resttemplate.fluent.get.GetMethod
@@ -15,11 +16,13 @@ import com.ofg.infrastructure.web.resttemplate.fluent.put.PutMethodBuilder
 import groovy.transform.CompileStatic
 import org.springframework.web.client.RestOperations
 
+import static com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders.NO_PREDEFINED_HEADERS
+
 /**
  * Point of entry of the fluent API over {@link RestOperations}.
  * This class gives methods for each of the HttpMethods and delegates to the root of
  * the fluent API of that method.
- * 
+ *
  * @see DeleteMethod
  * @see GetMethod
  * @see HeadMethod
@@ -36,38 +39,40 @@ class HttpMethodBuilder {
      * URL of an external URL or a service retrieved via service discovery 
      */
     private final String serviceUrl
+    private final PredefinedHttpHeaders predefinedHeaders
 
     HttpMethodBuilder(RestOperations restOperations) {
-        this('', restOperations)
+        this('', restOperations, NO_PREDEFINED_HEADERS)
     }
 
-    HttpMethodBuilder(String serviceUrl, RestOperations restOperations) {
+    HttpMethodBuilder(String serviceUrl, RestOperations restOperations, PredefinedHttpHeaders predefinedHeaders) {
+        this.predefinedHeaders = predefinedHeaders
         this.restOperations = restOperations
         this.serviceUrl = serviceUrl
     }
 
     DeleteMethod delete() {
-        return new DeleteMethodBuilder(serviceUrl, restOperations)
+        return new DeleteMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
     GetMethod get() {
-        return new GetMethodBuilder(serviceUrl, restOperations)
+        return new GetMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
     HeadMethod head() {
-        return new HeadMethodBuilder(serviceUrl, restOperations)
+        return new HeadMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
     OptionsMethod options() {
-        return new OptionsMethodBuilder(serviceUrl, restOperations)
+        return new OptionsMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
     PostMethod post() {
-        return new PostMethodBuilder(serviceUrl, restOperations)
+        return new PostMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
     PutMethod put() {
-        return new PutMethodBuilder(serviceUrl, restOperations)
+        return new PutMethodBuilder(serviceUrl, restOperations, predefinedHeaders)
     }
 
 }