+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/resttemplate/fluent/delete/DeleteMethodBuilder.groovy
@@ -2,10 +2,13 @@ package com.ofg.infrastructure.web.resttemplate.fluent.delete
 
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.BodylessWithHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.HeadersHaving
+import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders
 import groovy.transform.TypeChecked
 import org.springframework.http.HttpEntity
 import org.springframework.http.ResponseEntity
 import org.springframework.web.client.RestOperations
+
+import static com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders.NO_PREDEFINED_HEADERS
 
 /**
  * Implementation of the {@link org.springframework.http.HttpMethod#DELETE method} fluent API
@@ -19,14 +22,14 @@ class DeleteMethodBuilder implements DeleteMethod, UrlParameterizableDeleteMetho
     private final RestOperations restOperations
     @Delegate private final BodylessWithHeaders<ResponseReceivingDeleteMethod> withHeaders
 
-    DeleteMethodBuilder(String host, RestOperations restOperations) {
+    DeleteMethodBuilder(String host, RestOperations restOperations, PredefinedHttpHeaders predefinedHeaders) {
         this.restOperations = restOperations
         params.host = host
-        withHeaders =  new BodylessWithHeaders<ResponseReceivingDeleteMethod>(this, params)
+        withHeaders =  new BodylessWithHeaders<ResponseReceivingDeleteMethod>(this, params, predefinedHeaders)
     }
 
     DeleteMethodBuilder(RestOperations restOperations) {
-        this(EMPTY_HOST, restOperations)
+        this(EMPTY_HOST, restOperations, NO_PREDEFINED_HEADERS)
     }
 
     @Override