+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/resttemplate/fluent/get/GetMethodBuilder.groovy
@@ -3,11 +3,14 @@ package com.ofg.infrastructure.web.resttemplate.fluent.get
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.BodyContainingWithHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.HeadersHaving
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.ObjectReceiving
+import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.ResponseEntityReceiving
 import groovy.transform.TypeChecked
 import org.springframework.http.HttpEntity
 import org.springframework.http.ResponseEntity
 import org.springframework.web.client.RestOperations
+
+import static com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders.NO_PREDEFINED_HEADERS
 
 /**
  * Implementation of the {@link org.springframework.http.HttpMethod#GET method} fluent API
@@ -21,14 +24,14 @@ class GetMethodBuilder implements GetMethod, UrlParameterizableGetMethod, Respon
     private final RestOperations restOperations
     @Delegate private final BodyContainingWithHeaders withHeaders
 
-    GetMethodBuilder(String host, RestOperations restOperations) {
-        this.restOperations = restOperations
-        params.host = host
-        withHeaders = new BodyContainingWithHeaders(this, params)
+    GetMethodBuilder(RestOperations restOperations) {
+        this(EMPTY_HOST, restOperations, NO_PREDEFINED_HEADERS)
     }
 
-    GetMethodBuilder(RestOperations restOperations) {
-        this(EMPTY_HOST, restOperations)
+    GetMethodBuilder(String host, RestOperations restOperations, PredefinedHttpHeaders predefinedHeaders) {
+        this.restOperations = restOperations
+        params.host = host
+        withHeaders = new BodyContainingWithHeaders(this, params, predefinedHeaders)
     }
 
     @Override