+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/resttemplate/fluent/post/PostMethodBuilder.groovy
@@ -1,8 +1,10 @@
 package com.ofg.infrastructure.web.resttemplate.fluent.post
+
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.executor.LocationFindingExecutor
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.BodyContainingWithHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.HeadersSetting
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.ObjectReceiving
+import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders
 import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.ResponseEntityReceiving
 import groovy.transform.TypeChecked
 import org.springframework.http.HttpEntity
@@ -11,6 +13,7 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.client.RestOperations
 
 import static org.springframework.http.HttpMethod.POST
+import static com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders.NO_PREDEFINED_HEADERS
 
 /**
  * Implementation of the {@link org.springframework.http.HttpMethod#POST method} fluent API
@@ -24,14 +27,14 @@ class PostMethodBuilder extends LocationFindingExecutor implements
     
     @Delegate private final BodyContainingWithHeaders withHeaders
 
-    PostMethodBuilder(String host, RestOperations restOperations) {
+    PostMethodBuilder(String host, RestOperations restOperations, PredefinedHttpHeaders predefinedHeaders) {
         super(restOperations)
         params.host = host
-        withHeaders = new BodyContainingWithHeaders(this, params)
+        withHeaders = new BodyContainingWithHeaders(this, params, predefinedHeaders)
     }
     
     PostMethodBuilder(RestOperations restOperations) {
-        this(EMPTY_HOST, restOperations)
+        this(EMPTY_HOST, restOperations, NO_PREDEFINED_HEADERS)
     }
 
     @Override