+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/web/resttemplate/fluent/ServiceRestClientSpec.groovy
@@ -1,5 +1,6 @@
 package com.ofg.infrastructure.web.resttemplate.fluent
 
+import com.ofg.infrastructure.discovery.ServiceConfigurationResolver
 import com.ofg.infrastructure.discovery.ServiceResolver
 import com.ofg.infrastructure.discovery.ServiceUnavailableException
 import org.springframework.http.HttpEntity
@@ -13,10 +14,17 @@ class ServiceRestClientSpec extends Specification {
     public static final String COLA_COLLABORATOR_NAME = 'cola'
     RestOperations restOperations = Mock()
     ServiceResolver serviceResolver = Mock()
-    
-    ServiceRestClient serviceRestClient = new ServiceRestClient(restOperations, serviceResolver)
-    
-    def "should send a request to provided URL with appending host when calling service"() {
+    ServiceConfigurationResolver configurationResolver = Mock()
+
+    ServiceRestClient serviceRestClient = new ServiceRestClient(restOperations, serviceResolver, configurationResolver)
+
+    def setup() {
+        configurationResolver.getDependencies() >> ['cola': ['contentTypeTemplate': 'application/vnd.cola.$version+json',
+                                                             'version'            : 'v1',
+                                                             'headers'            : ['header1': 'value1', 'header2': 'value2']]]
+    }
+
+    def 'should send a request to provided URL with appending host when calling service'() {
         given:
             String serviceUrl = 'http://localhost:1234'
             String path = 'some/serviceUrl'
@@ -28,10 +36,55 @@ class ServiceRestClientSpec extends Specification {
         then:
             1 * restOperations.exchange(expectedUri, GET, _ as HttpEntity, _ as Class)
     }
-    
-    def "should throw an exception when trying to access an unavailable service"() {
+
+    def 'should send a request to provided URL with Content-Type set when calling service'() {
         given:
-            serviceResolver.fetchUrl(COLA_COLLABORATOR_NAME) >> { throw new ServiceUnavailableException(COLA_COLLABORATOR_NAME) }
+            String serviceUrl = 'http://localhost:1234'
+            String path = 'some/serviceUrl'
+        and:
+            serviceResolver.fetchUrl(COLA_COLLABORATOR_NAME) >> serviceUrl
+        when:
+            serviceRestClient.forService(COLA_COLLABORATOR_NAME).get().onUrl(path).ignoringResponse()
+        then:
+            1 * restOperations.exchange(_ as URI, GET, {
+                it.getHeaders().getContentType().toString() == 'application/vnd.cola.v1+json'
+            } as HttpEntity, _ as Class)
+    }
+
+    def 'should throw exception on creating Content-Type header when version property is required and is missing in dependency configuration'() {
+        given:
+            String serviceUrl = 'http://localhost:1234'
+            String path = 'some/serviceUrl'
+        and:
+            serviceResolver.fetchUrl(COLA_COLLABORATOR_NAME) >> serviceUrl
+        when:
+            serviceRestClient.forService(COLA_COLLABORATOR_NAME).get().onUrl(path).ignoringResponse()
+        then:
+            1 * configurationResolver.getDependencies() >> ['cola': ['contentTypeTemplate': 'application/vnd.cola.$version+json',
+                                                                     'headers'            : ['header1': 'value1', 'header2': 'value2']]]
+            thrown(MissingPropertyException)
+    }
+
+    def 'should send a request to provided URL with predefined headers set when calling service'() {
+        given:
+            String serviceUrl = 'http://localhost:1234'
+            String path = 'some/serviceUrl'
+        and:
+            serviceResolver.fetchUrl(COLA_COLLABORATOR_NAME) >> serviceUrl
+        when:
+            serviceRestClient.forService(COLA_COLLABORATOR_NAME).get().onUrl(path).ignoringResponse()
+        then:
+            1 * restOperations.exchange(_ as URI, GET, {
+                it.getHeaders().get('header1') == ['value1'] as List &&
+                        it.getHeaders().get('header2') == ['value2'] as List
+            } as HttpEntity, _ as Class)
+    }
+
+    def 'should throw an exception when trying to access an unavailable service'() {
+        given:
+            serviceResolver.fetchUrl(COLA_COLLABORATOR_NAME) >> {
+                throw new ServiceUnavailableException(COLA_COLLABORATOR_NAME)
+            }
         when:
             serviceRestClient.forService(COLA_COLLABORATOR_NAME).get().onUrl('').ignoringResponse()
         then: