+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/web/resttemplate/fluent/common/HttpMethodSpec.groovy
@@ -17,5 +17,5 @@ class HttpMethodSpec extends Specification {
 
     RestOperations restOperations = Mock()
     HttpMethodBuilder httpMethodBuilder
-    
+
 }