+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/web/resttemplate/fluent/headers/HeadersSettingSpec.groovy
@@ -2,6 +2,7 @@ package com.ofg.infrastructure.web.resttemplate.fluent.headers
 
 import com.ofg.infrastructure.web.resttemplate.fluent.HttpMethodBuilder
 import com.ofg.infrastructure.web.resttemplate.fluent.common.HttpMethodSpec
+import com.ofg.infrastructure.web.resttemplate.fluent.common.response.receive.PredefinedHttpHeaders
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.MediaType
@@ -12,7 +13,10 @@ import static org.springframework.http.MediaType.APPLICATION_JSON
 class HeadersSettingSpec extends HttpMethodSpec {
 
     public static final String TEMPLATE_URL = 'http://some.url/api/objects/{objectId}'
-    
+    private static final Map ADDITIONAL_HEADERS_CONFIG = ['headers': ['header1': 'value1', 'header2': 'value2']]
+    private static final Map CONTENT_TYPE_HEADER_CONFIG = ['contentTypeTemplate': 'application/vnd.external-service.$version+json', 'version': 'v1']
+    private static final String EXPECTED_CONTENT_TYPE = 'application/vnd.external-service.v1+json'
+
     def "should fill out headers for get method"() {
         given:
             httpMethodBuilder = new HttpMethodBuilder(restOperations)
@@ -97,7 +101,6 @@ class HeadersSettingSpec extends HttpMethodSpec {
                 _ as Long)
     }
 
-
     def "should fill out headers for options method"() {
         given:
             httpMethodBuilder = new HttpMethodBuilder(restOperations)
@@ -153,7 +156,6 @@ class HeadersSettingSpec extends HttpMethodSpec {
                     _ as Long)
     }
 
-
     def "should send a request from passed HttpEntity"() {
         given:
             String body = '''{"some":"body"}'''
@@ -172,6 +174,42 @@ class HeadersSettingSpec extends HttpMethodSpec {
                         httpEntity.headers.keySet().every(isPresentInSetHeaders()) && httpEntity.body == body } as HttpEntity,
                     _ as Class,
                     _ as Long)
+    }
+
+    def "should add custom headers to request when predefined headers provided"() {
+        given:
+            PredefinedHttpHeaders predefinedHttpHeaders = new PredefinedHttpHeaders(ADDITIONAL_HEADERS_CONFIG)
+            httpMethodBuilder = new HttpMethodBuilder(SERVICE_URL, restOperations, predefinedHttpHeaders)
+        when:
+            httpMethodBuilder
+                .get()
+                    .onUrlFromTemplate(URL_TEMPLATE)
+                        .withVariables(OBJECT_ID)
+                    .andExecuteFor()
+                        .anObject()
+                        .ofType(BigDecimal)
+        then:
+            1 * restOperations.exchange(FULL_URL, GET, {
+                HttpEntity httpEntity -> hasExpectedAdditionalHeaders(httpEntity)
+            } as HttpEntity, BigDecimal, OBJECT_ID)
+    }
+
+    def "should set Content-Type header when template with version is provided"() {
+        given:
+            PredefinedHttpHeaders predefinedHttpHeaders = new PredefinedHttpHeaders(CONTENT_TYPE_HEADER_CONFIG)
+            httpMethodBuilder = new HttpMethodBuilder(SERVICE_URL, restOperations, predefinedHttpHeaders)
+        when:
+            httpMethodBuilder
+                .get()
+                    .onUrlFromTemplate(URL_TEMPLATE)
+                       .withVariables(OBJECT_ID)
+                    .andExecuteFor()
+                        .anObject()
+                        .ofType(BigDecimal)
+        then:
+            1 * restOperations.exchange(FULL_URL, GET, {
+                HttpEntity httpEntity -> hasExpectedContentTypeHeader(httpEntity)
+            } as HttpEntity, BigDecimal, OBJECT_ID)
     }
 
     def "should fill out content type from media type"() {
@@ -278,6 +316,15 @@ class HeadersSettingSpec extends HttpMethodSpec {
 
     public final Closure<Boolean> isPresentInSetHeaders() {
         return { it in ['key', 'anotherKey', 'Accept', 'Cache-Control', 'Content-Type', 'Expires', 'Last-Modified', 'Location'] }
+    }
+
+    private boolean hasExpectedContentTypeHeader(HttpEntity entity) {
+        return entity.getHeaders().getContentType().toString() == EXPECTED_CONTENT_TYPE
+    }
+
+    private boolean hasExpectedAdditionalHeaders(HttpEntity entity) {
+        return entity.getHeaders().get('header1') == ['value1'] as List &&
+                entity.getHeaders().get('header2') == ['value2'] as List
     }
 
 }